Add option for synchronizing with a log file
To avoid a large download, we can save messages to a log and only fetch
messages newer than the newest file in the log. This adds basic support
but there is an inefficient step. Due to closing brackets in the JSON
format, it is not possible to keep the log updated by appending lines.
We have to sort and overwrite.

Signed-off-by: Connor Behan <connor.behan@gmail.com>